fixed some metrics serialization issues, fixed caching

+++ src/CMon/Models/Ccu/StateAndEventsResult.cs
@@ -4,7 +4,7 @@
 	{
 		public Input[] Inputs { get; set; }
 
-		public byte[] Outputs { get; set; }
+		public int[] Outputs { get; set; }
 
 		public string[] Partitions { get; set; }
